I had to kick the measure (make and revert change) to get version id and date modfied to update
on last commit changing min version in xml doesn't trigger update ot see the change, need to make change outside fo the xml. There end up being re-ording if files list in the XML not sure why that is the case. @nllong this is another odd behavior that results out of sync XMLs.